+++ examples/todo/lib/bloc/app.dart
@@ -1,3 +1,0 @@
-import 'package:bloc_lite/bloc_lite.dart';
-
-class AppController extends BlocController {}