+++ packages/bloc_lite/lib/bloc_lite.dart
@@ -1,5 +1,3 @@
 library bloc_lite;
 
 export './src/bloc_controller.dart';
-export './src/bloc_state_controller.dart';
-export './src/bloc_state.dart';