+++ packages/bloc_lite_flutter/lib/bloc_lite_flutter.dart
@@ -1,7 +1,7 @@
 library bloc_lite_flutter;
 
 export './src/typedefs.dart';
-export './src/bloc_state_widget.dart';
+export './src/bloc_builder.dart';
 export './src/bloc_widget.dart';
 export './src/inherited_bloc_tree.dart';
 export './src/inherited_bloc.dart';