+++ packages/bloc_lite_flutter/lib/src/inherited_bloc.dart
@@ -6,11 +6,12 @@ Type _typeOf<T>() => T;
 /// A Flutter widget which injects a [BlocController] into the widget tree.
 /// The bloc can then be retrieved as a dependency injection by multiple widget
 /// descendants within the child widget subtree.
-class InheritedBloc<B extends BlocController> extends InheritedWidget {
+class InheritedBloc<B extends BlocController> extends StatefulWidget {
   InheritedBloc({
     Key key,
     this.child,
     @required this.bloc,
+    this.autoDispose = true,
   })  : assert(bloc != null),
         super(key: key);
 
@@ -21,15 +22,19 @@ class InheritedBloc<B extends BlocController> extends InheritedWidget {
   /// The [BlocController] to be injected into the widget tree.
   final B bloc;
 
+  /// If true, the controller will be disposed when this widget leaves the widget
+  /// tree and is destroyed.
+  final bool autoDispose;
+
   /// A method which finds a [BlocController] of the specified subtype in the
   /// ancestor widget tree. Returns `null` if no [BlocController] is found.
   ///
   /// This method runs in O(1) time.
   static B of<B extends BlocController>(BuildContext context) {
-    final type = _typeOf<InheritedBloc<B>>();
+    final type = _typeOf<_InheritedBlocInheritedWidget<B>>();
     final provider = context
         .ancestorInheritedElementForWidgetOfExactType(type)
-        ?.widget as InheritedBloc<B>;
+        ?.widget as _InheritedBlocInheritedWidget<B>;
     return provider?.bloc;
   }
 
@@ -63,8 +68,8 @@ class InheritedBloc<B extends BlocController> extends InheritedWidget {
   static Iterable all(BuildContext context) sync* {
     final blocs = <BlocController>[];
     context.visitAncestorElements((elem) {
-      if (elem.widget is InheritedBloc) {
-        blocs.add((elem.widget as InheritedBloc).bloc);
+      if (elem.widget is _InheritedBlocInheritedWidget) {
+        blocs.add((elem.widget as _InheritedBlocInheritedWidget).bloc);
       }
       return true;
     });
@@ -80,5 +85,38 @@ class InheritedBloc<B extends BlocController> extends InheritedWidget {
   }
 
   @override
-  bool updateShouldNotify(InheritedBloc oldWidget) => false;
+  State createState() => _InheritedBlocState<B>();
+
+}
+
+class _InheritedBlocState<B extends BlocController> extends State<InheritedBloc<B>> {
+  @override
+  void dispose() {
+    if (widget.autoDispose) {
+      widget.bloc.dispose();
+    }
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext cxt) {
+    return _InheritedBlocInheritedWidget<B>(
+      bloc: widget.bloc,
+      child: widget.child,
+    );
+  }
+}
+
+class _InheritedBlocInheritedWidget<B extends BlocController> extends InheritedWidget {
+  _InheritedBlocInheritedWidget({
+    Key key,
+    this.bloc,
+    Widget child,
+  }) : super(key: key, child: child);
+
+  final B bloc;
+
+  @override
+  bool updateShouldNotify(InheritedWidget oldWidget) => false;
+
 }